# perf: use `indexOf()` intead of `includes()`